+++ examples/ExternalViewEngine/ExternalViewEngine.fs
@@ -69,13 +69,12 @@ let main args =
     bldr.Services
         .AddSingleton<ITemplate, ScribanTemplate>()
         // ^-- register ITemplates implementation as a dependency
-
         |> ignore
+
+    let wapp = bldr.Build()
 
     let endpoints =
         [ get "/" Pages.homepage ]
-
-    let wapp = bldr.Build()
 
     wapp.UseRouting()
         .UseFalco(endpoints)