+++ site/Site.fs
@@ -26,7 +26,6 @@ module Markdown =
             MarkdownPipelineBuilder()
                 .UseAutoIdentifiers()
                 .UsePipeTables()
-                .UseYamlFrontMatter()
                 .UseAutoLinks()
                 .Build()
 
@@ -54,7 +53,7 @@ module Markdown =
 
 
         // Rewrite direct markdown doc links
-        let body = Regex.Replace(renderedMarkdown, "([a-zA-Z\-]+)\.md", "$1.html")
+        let body = Regex.Replace(renderedMarkdown.Replace("\"documentation/", "\"docs/"), "([a-zA-Z\-]+)\.md", "$1.html")
 
         { Title = title |> Option.defaultValue ""
           Body = body }
@@ -67,11 +66,11 @@ module View =
         [
             Elem.h3 [] [ Text.raw "Project Links" ]
             Elem.a [ Attr.href "/"] [ Text.raw "Project Homepage" ]
-            Elem.a [ Attr.class' "db"; Attr.href "https://github.com/pimbrouwers/Falco"; Attr.targetBlank ]
+            Elem.a [ Attr.class' "db"; Attr.href "https://github.com/FalcoFramework/Falco"; Attr.targetBlank ]
                 [ Text.raw "Source Code" ]
-            Elem.a [ Attr.class' "db"; Attr.href "https://github.com/pimbrouwers/Falco/issues"; Attr.targetBlank ]
+            Elem.a [ Attr.class' "db"; Attr.href "https://github.com/FalcoFramework/Falco/issues"; Attr.targetBlank ]
                 [ Text.raw "Issue Tracker" ]
-            Elem.a [ Attr.class' "db"; Attr.href "https://github.com/pimbrouwers/Falco/discussions"; Attr.targetBlank ]
+            Elem.a [ Attr.class' "db"; Attr.href "https://github.com/FalcoFramework/Falco/discussions"; Attr.targetBlank ]
                 [ Text.raw "Discussion" ]
             Elem.a [ Attr.class' "db"; Attr.href "https://twitter.com/falco_framework"; Attr.targetBlank ]
                 [ Text.raw "Twitter" ]
@@ -82,27 +81,29 @@ module View =
             Text.h3 "Contents"
             Elem.ul [ Attr.class' "nl3 f6" ] [
                 Elem.li [] [ Elem.a [ Attr.href "get-started.html" ] [ Text.raw "Getting Started" ] ]
-                Elem.li [] [ Elem.a [ Attr.href "host-configuration.html" ] [ Text.raw "Host Configuration" ] ]
                 Elem.li [] [ Elem.a [ Attr.href "routing.html" ] [ Text.raw "Routing" ] ]
-                Elem.li [] [ Elem.a [ Attr.href "response.html" ] [ Text.raw "Writing responses" ] ]
-                Elem.li [] [ Elem.a [ Attr.href "request.html" ] [ Text.raw "Accessing request data" ] ]
-                Elem.li [] [ Elem.a [ Attr.href "markup.html" ] [ Text.raw "View engine" ] ]
+                Elem.li [] [ Elem.a [ Attr.href "response.html" ] [ Text.raw "Response Writing" ] ]
+                Elem.li [] [ Elem.a [ Attr.href "request.html" ] [ Text.raw "Request Handling" ] ]
+                Elem.li [] [ Elem.a [ Attr.href "markup.html" ] [ Text.raw "Markup" ] ]
                 Elem.li [] [
-                    Elem.a [ Attr.href "cross-site-request-forgery.html" ] [ Text.raw "Security" ]
+                    Text.raw "Security"
                     Elem.ul [] [
                         Elem.li [] [ Elem.a [ Attr.href "cross-site-request-forgery.html" ] [ Text.raw "Cross Site Request Forgery (XSRF)" ] ]
                         Elem.li [] [ Elem.a [ Attr.href "authentication.html" ] [ Text.raw "Authentication & Authorization" ] ]
                     ]
                 ]
+                Elem.li [] [ Elem.a [ Attr.href "host-configuration.html" ] [ Text.raw "Host Configuration" ] ]
                 Elem.li [] [ Elem.a [ Attr.href "deployment.html" ] [ Text.raw "Deployment" ] ]
                 Elem.li [] [
-                    Elem.a [ Attr.href "example-hello-world.html" ] [ Text.raw "Examples" ]
+                    Text.raw "Examples"
                     Elem.ul [] [
                         Elem.li [] [ Elem.a [ Attr.href "example-hello-world.html" ] [ Text.raw "Hello World" ] ]
                         Elem.li [] [ Elem.a [ Attr.href "example-hello-world-mvc.html" ] [ Text.raw "Hello World MVC" ] ]
                         Elem.li [] [ Elem.a [ Attr.href "example-dependency-injection.html" ] [ Text.raw "Dependency Injection" ] ]
-                        Elem.li [] [ Elem.a [ Attr.href "example-external-view-engine.html" ] [ Text.raw "Hello World" ] ]
+                        Elem.li [] [ Elem.a [ Attr.href "example-external-view-engine.html" ] [ Text.raw "External View Engine" ] ]
                         Elem.li [] [ Elem.a [ Attr.href "example-basic-rest-api.html" ] [ Text.raw "Basic REST API" ] ]
+                        Elem.li [] [ Elem.a [ Attr.href "example-open-api.html" ] [ Text.raw "Open API" ] ]
+                        Elem.li [] [ Elem.a [ Attr.href "example-htmx.html" ] [ Text.raw "htmx" ] ]
                     ]
                 ]
                 Elem.li [] [ Elem.a [ Attr.href "migrating-from-v4-to-v5.html" ] [ Text.raw "V5 Migration Guide" ] ]
@@ -154,11 +155,11 @@ module View =
                     Elem.div [ Attr.class' "flex-grow-1-l tc tr-l" ] [
                         Elem.a [ Attr.href "/docs"; Attr.title "Overview of Falco's key features"; Attr.class' "dib mh2 mh3-l no-underline white-90 hover-white" ]
                             [ Text.raw "docs" ]
-                        Elem.a [ Attr.href "https://github.com/pimbrouwers/Falco"; Attr.title "Fork Falco on GitHub"; Attr.alt "Falco GitHub Link"; Attr.targetBlank; Attr.class' "dib mh2 ml3-l no-underline white-90 hover-white" ]
+                        Elem.a [ Attr.href "https://github.com/FalcoFramework/Falco"; Attr.title "Fork Falco on GitHub"; Attr.alt "Falco GitHub Link"; Attr.targetBlank; Attr.class' "dib mh2 ml3-l no-underline white-90 hover-white" ]
                             [ Text.raw "code" ]
-                        Elem.a [ Attr.href "https://github.com/pimbrouwers/Falco/tree/master/examples"; Attr.title "Falco code samples"; Attr.alt "Faclo code samples link"; Attr.class' "dib ml2 mh3-l no-underline white-90 hover-white" ]
+                        Elem.a [ Attr.href "https://github.com/FalcoFramework/Falco/tree/master/examples"; Attr.title "Falco code samples"; Attr.alt "Faclo code samples link"; Attr.class' "dib ml2 mh3-l no-underline white-90 hover-white" ]
                             [ Text.raw "samples" ]
-                        Elem.a [ Attr.href "https://github.com/pimbrouwers/Falco/discussions"; Attr.title "Need help?"; Attr.alt "Faclo GitHub discussions link"; Attr.class' "dib ml2 mh3-l no-underline white-90 hover-white" ]
+                        Elem.a [ Attr.href "https://github.com/FalcoFramework/Falco/discussions"; Attr.title "Need help?"; Attr.alt "Faclo GitHub discussions link"; Attr.class' "dib ml2 mh3-l no-underline white-90 hover-white" ]
                             [ Text.raw "help" ]
                     ]
                 ]
@@ -217,7 +218,7 @@ module View =
                     Elem.img [ Attr.src "/icons/integrate.svg"; Attr.class' "w4 o-90" ]
                     Elem.h3 [ Attr.class' "mv2 white" ] [ Text.raw "Extensible" ]
                     Elem.div [ Attr.class' "mb3 white-90" ] [ Text.raw "Seamlessly integrates with existing libraries." ]
-                    Elem.a [ Attr.href "https://github.com/pimbrouwers/Falco/tree/master/samples/ScribanExample"; Attr.targetBlank; Attr.title "Example of incorporating a third-party view engine"; Attr.class' "dib mh2 pa2 f6 white ba b--white br2 no-underline" ]
+                    Elem.a [ Attr.href "https://github.com/FalcoFramework/Falco/tree/master/samples/ScribanExample"; Attr.targetBlank; Attr.title "Example of incorporating a third-party view engine"; Attr.class' "dib mh2 pa2 f6 white ba b--white br2 no-underline" ]
                         [ Text.raw "Explore How" ]
                  ]
             ]